folder creation and file upload to folder

+++ client/src/component/maincontainer/FileItem.js
@@ -1,15 +1,28 @@
 import React from "react";
 import Moment from "react-moment";
 import { deleteFileById } from "../../actions/file";
+import { selectFolder } from "../../actions/folderData";
 import { connect } from "react-redux";
+import file from "../../reducers/file";
 
-const FileItem = ({ file: { FileName, CreatedAt, Id }, deleteFileById }) => {
+const FileItem = ({ file, deleteFileById, selectFolder }) => {
+  var { FileName, CreatedAt, Id, IsDir, ParentID } = file === null ? {} : file;
   const deleteFile = (e) => {
-    deleteFileById(Id);
+    if (window.confirm("Do you want to delete file " + FileName + "?")) {
+      deleteFileById(Id);
+    }
+  };
+  var open = () => {
+    if (IsDir) {
+      console.log("opening folder ", file);
+      selectFolder(file);
+    }
   };
   return (
     <tr class="filerow">
-      <td class="cell filename">{FileName}</td>
+      <td class="cell filename" onClick={open}>
+        {FileName}
+      </td>
       <td class="cell uploadeddate">
         <Moment format="DD/MM/YYYY">{CreatedAt}</Moment>
       </td>
@@ -27,5 +40,5 @@ const FileItem = ({ file: { FileName, CreatedAt, Id }, deleteFileById }) => {
     </tr>
   );
 };
-
-export default connect(null, { deleteFileById })(FileItem);
+// onClick={}
+export default connect(null, { deleteFileById, selectFolder })(FileItem);
